pyproject.toml
[tool.poetry]
name = "tuft-cam"
version = "0.1.0"
description = ""
authors = ["jantoreh <jantorehorn@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Pillow = "^8.0.1"

[tool.poetry.dev-dependencies]
numpy = "^1.19.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
